requirements.txt
python-binance
python-kucoin
poloniex
gdax
